ci: host build + tests on every push; and 32 KB back from the blob detector
The host build IS the firmware's test rig — VM, ISA, missions and the
colour path are platform-agnostic C/C++ that hal_host.cpp fakes the
hardware for. It stopped linking once and nobody noticed for weeks, so
it now runs on every push: the host demo (which links the whole runtime)
plus the 8 make targets that need no model or submodule. The Zephyr
board build stays local — multi-GB NCS toolchain plus the proprietary
add-on — gated by the pre-push hook as before.

The DK-A image was at 92.36% of RAM, with under 30 KB of headroom. The
colour detector's flood-fill stack was int[16384] while every value it
holds is a pixel index <= 16383: 32 KB of pure padding, and the largest
object in the build. uint16_t saves exactly 32768 B, changes nothing
observable (test-color, test-color-move, test-color-service all pass),
and takes the image to 86.10% — headroom from ~28 KB to ~61 KB.

+++ src/ai/color_detect.c
@@ -73,7 +73,12 @@ enum {
 
 static unsigned char g_mask[MAX_PIXELS];  // 0 = no colour, else 1 + class slot
 static unsigned char g_seen[MAX_PIXELS];
-static int g_stack[MAX_PIXELS];
+// The flood-fill stack holds PIXEL INDICES, so its values are bounded by
+// MAX_PIXELS (16383 max) and 16 bits are two bytes more than enough. As `int`
+// this single array was 64 KB — on a DK image already at 92% of RAM, the
+// largest object in the build and 32 KB of pure padding. Widening back to int
+// on pop is free.
+static uint16_t g_stack[MAX_PIXELS];
 
 tc_color_blob tc_detect_color_blob(const uint8_t *rgb, int width, int height) {
   tc_color_blob out;
